+++ bootique-jetty-metrics/src/test/java/io/bootique/jetty/metrics/JettyMetricsModuleProviderTest.java
@@ -1,0 +1,12 @@
+package io.bootique.jetty.metrics;
+
+import io.bootique.test.junit.BQModuleProviderChecker;
+import org.junit.Test;
+
+public class JettyMetricsModuleProviderTest {
+
+    @Test
+    public void testAutoLoad() {
+        BQModuleProviderChecker.testPresentInJar(JettyMetricsModuleProvider.class);
+    }
+}